requirements.txt
flask
flask-sqlalchemy
python-dotenv
gunicorn
bcrypt
